requirements.txt
PySide6~=6.9.2
QtAwesome~=1.4.0
vdf~=3.4
semver~=3.0.4
PyGithub~=2.8.1
python-dotenv~=1.1.1